+++ src/main/java/com/zipline/entity/publicItem/enums/Category.java
@@ -1,0 +1,7 @@
+package com.zipline.entity.publicItem.enums;
+
+public enum Category {
+    SALE,       // 매매
+    DEPOSIT,    // 전세
+    MONTHLY     // 월세
+}
