+++ src/main/java/com/zipline/entity/publicItem/enums/CrawlStatus.java
@@ -1,0 +1,9 @@
+package com.zipline.entity.publicItem.enums;
+
+public enum CrawlStatus {
+    NEW,
+    PENDING,
+    PROCESSING,
+    COMPLETED,
+    FAILED
+}
