+++ src/main/java/com/zipline/entity/publicItem/enums/MigrationStatus.java
@@ -1,0 +1,9 @@
+package com.zipline.entity.publicItem.enums;
+
+public enum MigrationStatus {
+    PENDING,
+    COMPLETED,
+    FAILED
+}
+
+
