+++ src/main/java/com/zipline/entity/publicItem/enums/Platform.java
@@ -1,0 +1,8 @@
+package com.zipline.entity.publicItem.enums;
+
+public enum Platform {
+    NAVER,
+    ZIGBANG,
+    DABANG
+}
+
